# backend/internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: 网络运维监控平台 API
  description: |
    网络运维监控平台后端API规范，基于RESTful设计风格。

    ## 认证方式
    - 所有API需要通过Header传递API Key：`Authorization: Bearer <token>`
    - 登录接口返回token

    ## 响应格式
    所有接口统一响应格式：
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: IT Manager Team
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
  - url: http://localhost/api
    description: 生产环境
...
paths:
...
  /auth/skip-password-change:
    post:
      tags:
        - 认证
      summary: 跳过改密 (非首次场景)
      description: |
        C7: 用户在 `/change-password` 页可点"跳过"调本接口。
        写 `password_set_at = NOW()` + 清 `must_change_password = false` + 写 audit_log。

        主人 7/02 决策: **首次登录强改密不可跳**。
        - `reason=first_login` → 400 (seed admin/admin123 这类必须改)
        - `reason=optional` (或不传) → 允许 (用户主动取消改密, 不算"首次")
      operationId: skipPasswordChange
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  enum: [first_login, optional]
                  description: first_login=必改拒绝; optional=允许跳 (默认)
                  example: optional
      responses:
        '200':
          description: 跳过成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          description: 首次登录必改, 拒绝跳过

  # ==================== 仪表盘 ====================
...
  /users/{id}:
    get:
      tags:
        - 用户管理
      summary: 获取用户详情
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ==================== 通知渠道 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string

    # ==================== 认证 ====================
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        nickname:
          type: string
        email:
          type: string
        role:
          type: string
          # 权威词表 = migrations/000001 的 roles.code + 000013 兜底值 user。
          # 旧值 operator/viewer 仅在存量库里可能见到，服务端会折叠为 ops_user/readonly。
          enum: [admin, ops_admin, ops_user, auditor, readonly, user]
        created_at:
          type: string
          format: date-time
...