# backend/internal/models/user.go
package models

import (
	"database/sql/driver"
	"encoding/json"
	"time"

	"github.com/google/uuid"
	"gorm.io/gorm"
)

// User 用户
type User struct {
	ID           uuid.UUID  `json:"id" gorm:"type:uuid;primary_key;default:gen_random_uuid()"`
	Username     string     `json:"username" gorm:"uniqueIndex;size:50;not null"`
	PasswordHash string     `json:"-" gorm:"size:255;not null"`
	Nickname     string     `json:"nickname" gorm:"size:100"`
	Email        string     `json:"email" gorm:"size:255"`
	Phone        string     `json:"phone" gorm:"size:20"`
	Avatar       string     `json:"avatar" gorm:"size:500"`
	Status       string     `json:"status" gorm:"size:20;default:active"` // active, inactive, locked
	DepartmentID *uuid.UUID `json:"department_id" gorm:"type:uuid"`
	// 取值 = migrations/000001 的 roles.code：admin / ops_admin / ops_user / auditor / readonly
	// （另有 000013 兜底值 user，等同 readonly）。权威口径与权限矩阵见 docs/FIX-PLAN-AUTHZ.md §3。
	Role               string         `json:"role" gorm:"size:20;default:user"`
	FailedLogin        int            `json:"failed_login" gorm:"default:0"`
	LockedUntil        *time.Time     `json:"locked_until"`
	LastLogin          *time.Time     `json:"last_login"`
	LastLoginIP        string         `json:"last_login_ip" gorm:"size:50"`
	MustChangePassword bool           `json:"must_change_password" gorm:"default:true"`
	PasswordSetAt      *time.Time     `json:"password_set_at"`
	CreatedAt          time.Time      `json:"created_at"`
	UpdatedAt          time.Time      `json:"updated_at"`
	DeletedAt          gorm.DeletedAt `json:"-" gorm:"index"`
}

func (u *User) TableName() string {
	return "users"
}

// BeforeCreate 创建前
func (u *User) BeforeCreate(tx *gorm.DB) error {
	if u.ID == uuid.Nil {
		u.ID = uuid.New()
	}
	return nil
}

// APIKey API 密钥
type APIKey struct {
	ID          uuid.UUID  `json:"id" gorm:"type:uuid;primary_key;default:gen_random_uuid()"`
	UserID      uuid.UUID  `json:"user_id" gorm:"type:uuid;not null"`
	Name        string     `json:"name" gorm:"size:100;not null"`
	KeyHash     string     `json:"-" gorm:"size:255;not null"`
	Prefix      string     `json:"prefix" gorm:"size:20;not null"`
	Permissions StringList `json:"permissions" gorm:"type:text;not null;default:'[]';serializer:json"`
	IPWhitelist StringList `json:"ip_whitelist" gorm:"type:text;not null;default:'[]';serializer:json"`
	RateLimit   int        `json:"rate_limit" gorm:"default:1000"`
	ExpiresAt   *time.Time `json:"expires_at"`
	LastUsedAt  *time.Time `json:"last_used_at"`
	Status      string     `json:"status" gorm:"size:20;default:active"`
	CreatedAt   time.Time  `json:"created_at"`
	UpdatedAt   time.Time  `json:"updated_at"`
}

// StringList 自定义类型实现 GORM JSON 序列化（PG/SQLite 通用）
type StringList []string

// Scan implements sql.Scanner
func (s *StringList) Scan(value interface{}) error {
	if value == nil {
		*s = nil
		return nil
	}
	var b []byte
	switch v := value.(type) {
	case []byte:
		b = v
	case string:
		b = []byte(v)
	default:
		*s = nil
		return nil
	}
	return json.Unmarshal(b, s)
}

// Value implements driver.Valuer
func (s StringList) Value() (driver.Value, error) {
	if s == nil {
		return "[]", nil
	}
	return json.Marshal(s)
}

func (a *APIKey) TableName() string {
	return "api_keys"
}

// AuditLog 审计日志
type AuditLog struct {
	ID         uuid.UUID  `json:"id" gorm:"type:uuid;primary_key;default:gen_random_uuid()"`
	UserID     *uuid.UUID `json:"user_id" gorm:"type:uuid"`
	Username   string     `json:"username" gorm:"size:100"`
	Action     string     `json:"action" gorm:"size:50;not null"`
	Resource   string     `json:"resource" gorm:"size:100"`
	ResourceID *uuid.UUID `json:"resource_id" gorm:"type:uuid"`
	Method     string     `json:"method" gorm:"size:10"`
	Path       string     `json:"path" gorm:"size:500"`
	IP         string     `json:"ip" gorm:"size:50"`
	UserAgent  string     `json:"user_agent" gorm:"size:500"`
	Status     int        `json:"status"` // 200, 400, 401, 403, 500
	ErrorMsg   string     `json:"error_msg" gorm:"size:1000"`
	RequestID  string     `json:"request_id" gorm:"size:50"`
	CreatedAt  time.Time  `json:"created_at"`
}

func (a *AuditLog) TableName() string {
	return "audit_logs"
}